refactor(instui-cli): simplify CLI logic
