+++ packages/instui-cli/lib/commands/version.js
@@ -21,12 +21,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+
 const { info } = require('@instructure/command-utils')
+const pjson = require('../../package.json')
 
 exports.command = 'version'
 exports.desc = 'Return the current version of instui-cli and exit.'
 
-exports.handler = (argv) => {
-  const pjson = require('../../package.json')
+exports.handler = () => {
   info(pjson.version)
 }
